refactor: Move ImageResizer to ui folder
