+++ packages/lexical-playground/src/nodes/ExcalidrawNode/index.jsx
@@ -26,7 +26,7 @@ import {
 import * as React from 'react';
 import {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 
-import {ImageResizer} from '../ImageNode';
+import ImageResizer from '../../ui/ImageResizer';
 import ExcalidrawImage from './ExcalidrawImage';
 import ExcalidrawModal from './ExcalidrawModal';
