+++ packages/lexical-playground/src/ui/ImageResizer.jsx
@@ -1,0 +1,182 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow strict
+ */
+
+import type {LexicalEditor} from 'lexical';
+
+import * as React from 'react';
+import {useRef} from 'react';
+
+export default function ImageResizer({
+  onResizeStart,
+  onResizeEnd,
+  imageRef,
+  editor,
+  showCaption,
+  setShowCaption,
+}: {
+  editor: LexicalEditor,
+  imageRef: {current: null | HTMLElement},
+  onResizeEnd: ('inherit' | number, 'inherit' | number) => void,
+  onResizeStart: () => void,
+  setShowCaption: (boolean) => void,
+  showCaption: boolean,
+}): React.Node {
+  const buttonRef = useRef(null);
+  const positioningRef = useRef<{
+    currentHeight: 'inherit' | number,
+    currentWidth: 'inherit' | number,
+    direction: 0 | 1 | 2 | 3,
+    isResizing: boolean,
+    ratio: number,
+    startHeight: number,
+    startWidth: number,
+    startX: number,
+    startY: number,
+  }>({
+    currentHeight: 0,
+    currentWidth: 0,
+    direction: 0,
+    isResizing: false,
+    ratio: 0,
+    startHeight: 0,
+    startWidth: 0,
+    startX: 0,
+    startY: 0,
+  });
+  const editorRootElement = editor.getRootElement();
+  // Find max width, accounting for editor padding.
+  const maxWidthContainer =
+    editorRootElement !== null
+      ? editorRootElement.getBoundingClientRect().width - 20
+      : 100;
+
+  const handlePointerDown = (event: PointerEvent, direction: 0 | 1 | 2 | 3) => {
+    const image = imageRef.current;
+    if (image !== null) {
+      const {width, height} = image.getBoundingClientRect();
+      const positioning = positioningRef.current;
+      positioning.startWidth = width;
+      positioning.startHeight = height;
+      positioning.ratio = width / height;
+      positioning.currentWidth = 'inherit';
+      positioning.currentHeight = 'inherit';
+      positioning.startX = event.clientX;
+      positioning.startY = event.clientY;
+      positioning.isResizing = true;
+      positioning.direction = direction;
+      onResizeStart();
+      document.addEventListener('pointermove', handlePointerMove);
+      document.addEventListener('pointerup', handlePointerUp);
+    }
+  };
+  const handlePointerMove = (event: PointerEvent) => {
+    const image = imageRef.current;
+    const positioning = positioningRef.current;
+
+    if (image !== null && positioning.isResizing) {
+      if (positioning.direction === 3) {
+        const diff = Math.floor(positioning.startY - event.clientY) * 2;
+        const minHeight = 150 * positioning.ratio;
+        const maxHeight = maxWidthContainer / positioning.ratio;
+        let height = positioning.startHeight + diff;
+        if (height < minHeight) {
+          height = minHeight;
+        } else if (height > maxHeight) {
+          height = maxHeight;
+        }
+        image.style.width = `inherit`;
+        image.style.height = `${height}px`;
+        positioning.currentHeight = height;
+      } else if (positioning.direction === 2) {
+        const diff = Math.floor(event.clientY - positioning.startY);
+        const minHeight = 150 * positioning.ratio;
+        const maxHeight = maxWidthContainer / positioning.ratio;
+        let height = positioning.startHeight + diff;
+        if (height < minHeight) {
+          height = minHeight;
+        } else if (height > maxHeight) {
+          height = maxHeight;
+        }
+        image.style.width = `inherit`;
+        image.style.height = `${height}px`;
+        positioning.currentHeight = height;
+      } else {
+        const diff = Math.floor(event.clientX - positioning.startX);
+        const minWidth = 150 * positioning.ratio;
+        const maxWidth = maxWidthContainer;
+        let width = positioning.startWidth + diff;
+        if (width < minWidth) {
+          width = minWidth;
+        } else if (width > maxWidth) {
+          width = maxWidth;
+        }
+        image.style.width = `${width}px`;
+        image.style.height = `inherit`;
+        positioning.currentWidth = width;
+      }
+    }
+  };
+  const handlePointerUp = (_event: PointerEvent) => {
+    const image = imageRef.current;
+    const positioning = positioningRef.current;
+    if (image !== null && positioning.isResizing) {
+      const width = positioning.currentWidth;
+      const height = positioning.currentHeight;
+      positioning.startWidth = 0;
+      positioning.startHeight = 0;
+      positioning.ratio = 0;
+      positioning.startX = 0;
+      positioning.startY = 0;
+      positioning.currentWidth = 0;
+      positioning.currentHeight = 0;
+      positioning.isResizing = false;
+      onResizeEnd(width, height);
+      document.removeEventListener('pointermove', handlePointerMove);
+      document.removeEventListener('pointerup', handlePointerUp);
+    }
+  };
+  return (
+    <>
+      {!showCaption && (
+        <button
+          className="image-caption-button"
+          ref={buttonRef}
+          onClick={() => {
+            setShowCaption(!showCaption);
+          }}>
+          Add Caption
+        </button>
+      )}
+      <div
+        className="image-resizer-ne"
+        onPointerDown={(event) => {
+          handlePointerDown(event, 0);
+        }}
+      />
+      <div
+        className="image-resizer-se"
+        onPointerDown={(event) => {
+          handlePointerDown(event, 1);
+        }}
+      />
+      <div
+        className="image-resizer-sw"
+        onPointerDown={(event) => {
+          handlePointerDown(event, 2);
+        }}
+      />
+      <div
+        className="image-resizer-nw"
+        onPointerDown={(event) => {
+          handlePointerDown(event, 3);
+        }}
+      />
+    </>
+  );
+}
